fix(security): sanitize provider values in logs, break the api import cycle
CodeQL findings on #555.

Log injection (4): `provider` reaches log lines straight from
AppConfig.ai_provider, which the user sets via PUT /api/config, so it is
untrusted on every one of them. str(e) on an httpx error embeds the request URL,
built from the equally user-settable `model`. Both now go through
sanitize_log_value, which is what the rest of the codebase already uses. The
earlier pass sanitized only the provider response body and missed that the
provider name itself is user input.

Cyclic import (1): validation.py imported require_localhost_or_lan from
routes.py, which imports validation.py back inside function bodies. That never
deadlocked at runtime, which is why it passed review, but it is a real cycle in
the import graph and puts a small validators module downstream of a 4000-line
route module. The network-origin gates move to a leaf module, app/api/guards.py,
that both routers import. routes.py re-exports the same function objects, so
every existing import and dependency_overrides key keeps working (verified by
identity).
