+++ backend/app/api/routes.py
@@ -2,7 +2,6 @@
 
 import asyncio
 import io
-import ipaddress
 import json
 import logging
 import os
@@ -26,6 +25,7 @@ from sqlalchemy import func
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlmodel import select
 
+from app.api.guards import is_loopback, require_localhost, require_localhost_or_lan
 from app.config import settings
 from app.core.discdb_exporter import get_makemkv_log_dir
 from app.core.errors import AIProviderError
@@ -74,78 +74,11 @@ def require_debug() -> None:
         raise HTTPException(status_code=403, detail="Simulation only available in debug mode")
 
 
-def _is_loopback(host: str | None) -> bool:
-    """True if ``host`` names the local machine.
-
-    Covers every loopback form a peer address can take — IPv4 (`127.0.0.0/8`),
-    IPv6 (`::1`), and the IPv4-mapped IPv6 loopback (`::ffff:127.0.0.1`) that
-    arrives on dual-stack binds (`HOST=::`). `localhost` is kept as an explicit
-    fallback: Starlette normally reports a numeric peer address, but a literal
-    hostname is accepted rather than silently rejected.
-    """
-    if not host:
-        return False
-    try:
-        addr = ipaddress.ip_address(host)
-    except ValueError:
-        return host == "localhost"
-    if addr.is_loopback:
-        return True
-    # Python < 3.13's is_loopback does not unwrap IPv4-mapped addresses, so
-    # check the mapped IPv4 explicitly for version-independent correctness.
-    mapped = getattr(addr, "ipv4_mapped", None)
-    return bool(mapped and mapped.is_loopback)
-
-
-def require_localhost(request: Request) -> None:
-    """FastAPI dependency: 403 unless the request came from the host machine.
-
-    Used by endpoints that surface or mutate privacy-sensitive local data
-    (e.g. ripping history, DiscDB/fingerprint contributions) so they stay
-    reachable from the dashboard but not from LAN peers when
-    `allow_lan_access=True` opens the bind to all interfaces. Tests should
-    override the dependency via
-    `app.dependency_overrides[require_localhost] = lambda: None` rather than
-    spoofing peer addresses.
-    """
-    if not _is_loopback(request.client.host if request.client else None):
-        raise HTTPException(
-            status_code=403, detail="This endpoint is only reachable from the host machine"
-        )
-
-
-async def require_localhost_or_lan(request: Request) -> None:
-    """FastAPI dependency: allow loopback always, LAN peers only when opted in.
-
-    Like `require_localhost`, but a non-loopback (LAN) client is also permitted
-    when the user has explicitly enabled `allow_lan_access`. Used by the manual
-    *import* endpoints so headless/Docker deployments — where the dashboard is
-    reached from another machine — can browse and import, while the more
-    sensitive fingerprint/contribution endpoints stay strictly host-only via
-    `require_localhost`.
-
-    The filesystem-browse surface this exposes is why the gate is the explicit
-    opt-in rather than always-on: enabling `allow_lan_access` is the user's
-    statement that they trust their network.
-    """
-    if _is_loopback(request.client.host if request.client else None):
-        return
-    try:
-        from app.services.config_service import get_config
-
-        config = await get_config()
-        allow_lan = bool(config.allow_lan_access)
-    except Exception:  # noqa: BLE001 — a config read failure must fail closed
-        allow_lan = False
-    if not allow_lan:
-        raise HTTPException(
-            status_code=403,
-            detail=(
-                "This endpoint is only reachable from the host machine. To use it "
-                "from another device (e.g. a Docker deployment), enable LAN access "
-                "in Settings."
-            ),
-        )
+# Network-origin gates live in app/api/guards.py so validation.py can depend on
+# them without importing this module (which imports validation.py back, creating
+# an import cycle). Re-exported here: many call sites and tests reference them as
+# app.api.routes.require_localhost, including dependency_overrides keys.
+_is_loopback = is_loopback
 
 
 # Request/Response Models
